Move pill inventory to gear tab
