+++ index.html
@@ -859,6 +859,7 @@
               <div class="gear-tabs">
                 <button class="gear-tab-btn active" data-tab="gearEquip">Gear</button>
                 <button class="gear-tab-btn" data-tab="gearAbilities">Abilities</button>
+                <button class="gear-tab-btn" data-tab="gearPills">Pills</button>
                 <button class="gear-tab-btn" data-tab="gearStats">Stats</button>
               </div>
             <div id="gearEquipSubTab" class="gear-tab-content active">
@@ -887,6 +888,20 @@
               <div id="gearAbilitiesSubTab" class="gear-tab-content" style="display:none;">
                 <div id="abilitySlots"></div>
                 <div id="availableAbilities"></div>
+              </div>
+              <div id="gearPillsSubTab" class="gear-tab-content" style="display:none;">
+                <div class="row">
+                  <div class="col">
+                    <h4>Temporary Effects</h4>
+                    <ul id="tempPillsList"></ul>
+                    <p class="muted" id="noTempPills" style="display:none;">No pills yet</p>
+                  </div>
+                  <div class="col">
+                    <h4>Permanent Effects</h4>
+                    <ul id="permPillsList"></ul>
+                    <p class="muted" id="noPermPills" style="display:none;">No pills yet</p>
+                  </div>
+                </div>
               </div>
               <div id="gearStatsSubTab" class="gear-tab-content" style="display:none;">
                 <div class="stat"><span>HP</span><span id="stat-hp">100/100</span></div>
@@ -1122,7 +1137,6 @@
           <button class="alchemy-subtab active" data-subtab="recipe-book">Recipe Book</button>
           <button class="alchemy-subtab" data-subtab="lab">Lab</button>
           <button class="alchemy-subtab" data-subtab="coalesce">Coalesce</button>
-          <button class="alchemy-subtab" data-subtab="pill-inventory">Pill Inventory</button>
         </div>
 
         <div id="recipe-book" class="alchemy-subtab-content">
@@ -1173,21 +1187,6 @@
         <div id="coalesce" class="alchemy-subtab-content" style="display:none;">
           <div class="cards">
             <div class="card"><p class="muted">Coalesce coming soon...</p></div>
-          </div>
-        </div>
-
-        <div id="pill-inventory" class="alchemy-subtab-content" style="display:none;">
-          <div class="cards">
-            <div class="card">
-              <h4>Temporary Pills</h4>
-              <ul id="tempPillsList"></ul>
-              <p class="muted" id="noTempPills" style="display:none;">No pills yet <button class="btn small ghost pill-open-recipes">Recipe Book</button></p>
-            </div>
-            <div class="card">
-              <h4>Permanent Pills</h4>
-              <ul id="permPillsList"></ul>
-              <p class="muted" id="noPermPills" style="display:none;">No pills yet <button class="btn small ghost pill-open-recipes">Recipe Book</button></p>
-            </div>
           </div>
         </div>
       </section>
